Randomly partitions data for cross-validation.

For a vector of training labels, we return a list of cross-validation folds, where each fold has the indices of the observations to leave out in the fold. In terms of classification error rate estimation, one can think of a fold as a the observations to hold out as a test sample set. Either the `hold_out` size or the number of folds, `num_folds`, can be specified. The number of folds defaults to 10, but if the `hold_out` size is specified, then `num_folds` is ignored.

cv_partition(y, num_folds = 10, hold_out = NULL, seed = NULL)

Arguments

y
a vector of class labels

num_folds
the number of cross-validation folds. Ignored if `hold_out` is not `NULL`. See Details.

hold_out
the hold-out size for cross-validation. See Details.

seed
optional random number seed for splitting the data for cross-validation

Value

list the indices of the training and test observations for each fold.

Details

We partition the vector `y` based on its length, which we treat as the sample size, 'n'. If an object other than a vector is used in `y`, its length can yield unexpected results. For example, the output of `length(diag(3))` is 9.

Examples

# The following three calls to `cv_partition` yield the same partitions.
set.seed(42)
cv_partition(iris$Species)

#> $Fold1
#> $Fold1$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  15  16  17  19  20
#>  [19]  21  22  23  24  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  41  42  43  44  45  46  47  48  50  51  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  73  74  75  76  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 101 102 103 104 105 106 108 110 111 112 113 114 115 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139
#> [127] 140 141 143 144 146 147 148 149 150
#> 
#> $Fold1$test
#>  [1]  49 100 109 142  18 145  40 107  77  72  14  25 116  52  78
#> 
#> 
#> $Fold2
#> $Fold2$training
#>   [1]   1   2   3   4   6   7   8   9  10  11  12  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  27  28  29  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  60  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  90  91  92  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 125 126 128 131 133 134 136 137 138 139 140
#> [127] 141 142 143 144 145 146 147 149 150
#> 
#> $Fold2$test
#>  [1]  65  89   5  30 130  26 129  93  13 135 132  66 127 148  59
#> 
#> 
#> $Fold3
#> $Fold3$training
#>   [1]   1   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  36  37  38  39
#>  [37]  40  41  42  44  45  46  47  48  49  50  51  52  54  55  56  57  58  59
#>  [55]  60  62  64  65  66  67  68  69  70  71  72  73  75  76  77  78  79  80
#>  [73]  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 107 108 109 111 112 113 114 115 116 117 119 121
#> [109] 122 123 124 125 127 128 129 130 131 132 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold3$test
#>  [1]  74 110  84  43 126   6  33 120  53 118  61  63 133  19 106
#> 
#> 
#> $Fold4
#> $Fold4$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  16  17  18  19  21
#>  [19]  22  23  24  25  26  27  29  30  31  32  33  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 126 127 128 129 130 131 132 133 134 135 136
#> [127] 139 141 142 143 144 145 147 148 150
#> 
#> $Fold4$test
#>  [1] 146  20  34  15  69 105 140 138  54 149  81 125 137  11  28
#> 
#> 
#> $Fold5
#> $Fold5$training
#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  23  24  25  26  27  28  29  30  31  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  74  75  76  77  78  79
#>  [73]  81  82  83  84  85  86  87  88  89  90  91  93  94  95  96  97  98  99
#>  [91] 100 102 104 105 106 107 108 109 110 111 112 113 115 116 118 119 120 121
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 144 145 146 148 149 150
#> 
#> $Fold5$test
#>  [1] 122 114  92  22   4   2  73  80  32  48 103 147 117  51 101
#> 
#> 
#> $Fold6
#> $Fold6$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57
#>  [55]  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  99 100 101 102 103 105 106 107 109 110 113 114 115 116 117 118
#> [109] 119 120 121 122 125 126 127 128 129 130 131 132 133 134 135 137 138 139
#> [127] 140 141 142 144 145 146 147 148 149
#> 
#> $Fold6$test
#>  [1] 150 111 104 123  98 124  29  55  60 136 108 143  79  37 112
#> 
#> 
#> $Fold7
#> $Fold7$training
#>   [1]   1   2   4   5   6   7   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  47  48  49  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  70  71  72  73  74  75  77  78  79  80
#>  [73]  81  82  83  84  86  87  88  89  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 129 130 132 133 135 136 137 138 139 140
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold7$test
#>  [1] 128 131   3   8  50  21  76  90  85  64 141  31  91  46 134
#> 
#> 
#> $Fold8
#> $Fold8$training
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  34  35  37  39  40  42
#>  [37]  43  44  45  46  48  49  50  51  52  53  54  55  56  57  59  60  61  63
#>  [55]  64  65  66  67  68  69  71  72  73  74  76  77  78  79  80  81  82  83
#>  [73]  84  85  86  87  88  89  90  91  92  93  95  97  98 100 101 102 103 104
#>  [91] 105 106 107 108 109 110 111 112 113 114 115 116 117 118 120 121 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold8$test
#>  [1]  47  41  58  36  99  96   9  94  17  38  62 119  23  70  75
#> 
#> 
#> $Fold9
#> $Fold9$training
#>   [1]   2   3   4   5   6   7   8   9  10  11  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  25  26  27  28  29  30  31  32  33  34  36  37  38  40  41
#>  [37]  42  43  46  47  48  49  50  51  52  53  54  55  56  58  59  60  61  62
#>  [55]  63  64  65  66  67  69  70  71  72  73  74  75  76  77  78  79  80  81
#>  [73]  82  83  84  85  86  87  89  90  91  92  93  94  96  98  99 100 101 103
#>  [91] 104 105 106 107 108 109 110 111 112 113 114 116 117 118 119 120 121 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold9$test
#>  [1]  24 139  97  68  88 115  35  57  44   1 102  39  12  45  95
#> 
#> 
#> $Fold10
#> $Fold10$training
#>   [1]   1   2   3   4   5   6   8   9  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  23  24  25  26  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  43  44  45  46  47  48  49  50  51  52  53  54  55  57  58  59  60
#>  [55]  61  62  63  64  65  66  68  69  70  72  73  74  75  76  77  78  79  80
#>  [73]  81  84  85  88  89  90  91  92  93  94  95  96  97  98  99 100 101 102
#>  [91] 103 104 105 106 107 108 109 110 111 112 114 115 116 117 118 119 120 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 145 146 147 148 149 150
#> 
#> $Fold10$test
#>  [1]  71  27  42  86  87  10  16 121  83 144  67 113  56  82   7
#> 
#> 

cv_partition(iris$Species, num_folds = 10, seed = 42)

#> $Fold1
#> $Fold1$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  15  16  17  19  20
#>  [19]  21  22  23  24  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  41  42  43  44  45  46  47  48  50  51  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  73  74  75  76  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 101 102 103 104 105 106 108 110 111 112 113 114 115 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139
#> [127] 140 141 143 144 146 147 148 149 150
#> 
#> $Fold1$test
#>  [1]  49 100 109 142  18 145  40 107  77  72  14  25 116  52  78
#> 
#> 
#> $Fold2
#> $Fold2$training
#>   [1]   1   2   3   4   6   7   8   9  10  11  12  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  27  28  29  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  60  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  90  91  92  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 125 126 128 131 133 134 136 137 138 139 140
#> [127] 141 142 143 144 145 146 147 149 150
#> 
#> $Fold2$test
#>  [1]  65  89   5  30 130  26 129  93  13 135 132  66 127 148  59
#> 
#> 
#> $Fold3
#> $Fold3$training
#>   [1]   1   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  36  37  38  39
#>  [37]  40  41  42  44  45  46  47  48  49  50  51  52  54  55  56  57  58  59
#>  [55]  60  62  64  65  66  67  68  69  70  71  72  73  75  76  77  78  79  80
#>  [73]  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 107 108 109 111 112 113 114 115 116 117 119 121
#> [109] 122 123 124 125 127 128 129 130 131 132 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold3$test
#>  [1]  74 110  84  43 126   6  33 120  53 118  61  63 133  19 106
#> 
#> 
#> $Fold4
#> $Fold4$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  16  17  18  19  21
#>  [19]  22  23  24  25  26  27  29  30  31  32  33  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 126 127 128 129 130 131 132 133 134 135 136
#> [127] 139 141 142 143 144 145 147 148 150
#> 
#> $Fold4$test
#>  [1] 146  20  34  15  69 105 140 138  54 149  81 125 137  11  28
#> 
#> 
#> $Fold5
#> $Fold5$training
#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  23  24  25  26  27  28  29  30  31  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  74  75  76  77  78  79
#>  [73]  81  82  83  84  85  86  87  88  89  90  91  93  94  95  96  97  98  99
#>  [91] 100 102 104 105 106 107 108 109 110 111 112 113 115 116 118 119 120 121
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 144 145 146 148 149 150
#> 
#> $Fold5$test
#>  [1] 122 114  92  22   4   2  73  80  32  48 103 147 117  51 101
#> 
#> 
#> $Fold6
#> $Fold6$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57
#>  [55]  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  99 100 101 102 103 105 106 107 109 110 113 114 115 116 117 118
#> [109] 119 120 121 122 125 126 127 128 129 130 131 132 133 134 135 137 138 139
#> [127] 140 141 142 144 145 146 147 148 149
#> 
#> $Fold6$test
#>  [1] 150 111 104 123  98 124  29  55  60 136 108 143  79  37 112
#> 
#> 
#> $Fold7
#> $Fold7$training
#>   [1]   1   2   4   5   6   7   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  47  48  49  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  70  71  72  73  74  75  77  78  79  80
#>  [73]  81  82  83  84  86  87  88  89  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 129 130 132 133 135 136 137 138 139 140
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold7$test
#>  [1] 128 131   3   8  50  21  76  90  85  64 141  31  91  46 134
#> 
#> 
#> $Fold8
#> $Fold8$training
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  34  35  37  39  40  42
#>  [37]  43  44  45  46  48  49  50  51  52  53  54  55  56  57  59  60  61  63
#>  [55]  64  65  66  67  68  69  71  72  73  74  76  77  78  79  80  81  82  83
#>  [73]  84  85  86  87  88  89  90  91  92  93  95  97  98 100 101 102 103 104
#>  [91] 105 106 107 108 109 110 111 112 113 114 115 116 117 118 120 121 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold8$test
#>  [1]  47  41  58  36  99  96   9  94  17  38  62 119  23  70  75
#> 
#> 
#> $Fold9
#> $Fold9$training
#>   [1]   2   3   4   5   6   7   8   9  10  11  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  25  26  27  28  29  30  31  32  33  34  36  37  38  40  41
#>  [37]  42  43  46  47  48  49  50  51  52  53  54  55  56  58  59  60  61  62
#>  [55]  63  64  65  66  67  69  70  71  72  73  74  75  76  77  78  79  80  81
#>  [73]  82  83  84  85  86  87  89  90  91  92  93  94  96  98  99 100 101 103
#>  [91] 104 105 106 107 108 109 110 111 112 113 114 116 117 118 119 120 121 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold9$test
#>  [1]  24 139  97  68  88 115  35  57  44   1 102  39  12  45  95
#> 
#> 
#> $Fold10
#> $Fold10$training
#>   [1]   1   2   3   4   5   6   8   9  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  23  24  25  26  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  43  44  45  46  47  48  49  50  51  52  53  54  55  57  58  59  60
#>  [55]  61  62  63  64  65  66  68  69  70  72  73  74  75  76  77  78  79  80
#>  [73]  81  84  85  88  89  90  91  92  93  94  95  96  97  98  99 100 101 102
#>  [91] 103 104 105 106 107 108 109 110 111 112 114 115 116 117 118 119 120 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 145 146 147 148 149 150
#> 
#> $Fold10$test
#>  [1]  71  27  42  86  87  10  16 121  83 144  67 113  56  82   7
#> 
#> 

cv_partition(iris$Species, hold_out = 15, seed = 42)

#> $Fold1
#> $Fold1$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  15  16  17  19  20
#>  [19]  21  22  23  24  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  41  42  43  44  45  46  47  48  50  51  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  73  74  75  76  79  80  81
#>  [73]  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 101 102 103 104 105 106 108 110 111 112 113 114 115 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139
#> [127] 140 141 143 144 146 147 148 149 150
#> 
#> $Fold1$test
#>  [1]  49 100 109 142  18 145  40 107  77  72  14  25 116  52  78
#> 
#> 
#> $Fold2
#> $Fold2$training
#>   [1]   1   2   3   4   6   7   8   9  10  11  12  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  27  28  29  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58
#>  [55]  60  61  62  63  64  67  68  69  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  90  91  92  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 125 126 128 131 133 134 136 137 138 139 140
#> [127] 141 142 143 144 145 146 147 149 150
#> 
#> $Fold2$test
#>  [1]  65  89   5  30 130  26 129  93  13 135 132  66 127 148  59
#> 
#> 
#> $Fold3
#> $Fold3$training
#>   [1]   1   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  36  37  38  39
#>  [37]  40  41  42  44  45  46  47  48  49  50  51  52  54  55  56  57  58  59
#>  [55]  60  62  64  65  66  67  68  69  70  71  72  73  75  76  77  78  79  80
#>  [73]  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 107 108 109 111 112 113 114 115 116 117 119 121
#> [109] 122 123 124 125 127 128 129 130 131 132 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold3$test
#>  [1]  74 110  84  43 126   6  33 120  53 118  61  63 133  19 106
#> 
#> 
#> $Fold4
#> $Fold4$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  13  14  16  17  18  19  21
#>  [19]  22  23  24  25  26  27  29  30  31  32  33  35  36  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  51  52  53  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  70  71  72  73  74  75  76  77  78  79
#>  [73]  80  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 100 101 102 103 104 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 120 121 122 123 124 126 127 128 129 130 131 132 133 134 135 136
#> [127] 139 141 142 143 144 145 147 148 150
#> 
#> $Fold4$test
#>  [1] 146  20  34  15  69 105 140 138  54 149  81 125 137  11  28
#> 
#> 
#> $Fold5
#> $Fold5$training
#>   [1]   1   3   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  21  23  24  25  26  27  28  29  30  31  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  69  70  71  72  74  75  76  77  78  79
#>  [73]  81  82  83  84  85  86  87  88  89  90  91  93  94  95  96  97  98  99
#>  [91] 100 102 104 105 106 107 108 109 110 111 112 113 115 116 118 119 120 121
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 144 145 146 148 149 150
#> 
#> $Fold5$test
#>  [1] 122 114  92  22   4   2  73  80  32  48 103 147 117  51 101
#> 
#> 
#> $Fold6
#> $Fold6$training
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  30  31  32  33  34  35  36  38
#>  [37]  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57
#>  [55]  58  59  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76
#>  [73]  77  78  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95
#>  [91]  96  97  99 100 101 102 103 105 106 107 109 110 113 114 115 116 117 118
#> [109] 119 120 121 122 125 126 127 128 129 130 131 132 133 134 135 137 138 139
#> [127] 140 141 142 144 145 146 147 148 149
#> 
#> $Fold6$test
#>  [1] 150 111 104 123  98 124  29  55  60 136 108 143  79  37 112
#> 
#> 
#> $Fold7
#> $Fold7$training
#>   [1]   1   2   4   5   6   7   9  10  11  12  13  14  15  16  17  18  19  20
#>  [19]  22  23  24  25  26  27  28  29  30  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  47  48  49  51  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  65  66  67  68  69  70  71  72  73  74  75  77  78  79  80
#>  [73]  81  82  83  84  86  87  88  89  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 120 121 122 123 124 125 126 127 129 130 132 133 135 136 137 138 139 140
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold7$test
#>  [1] 128 131   3   8  50  21  76  90  85  64 141  31  91  46 134
#> 
#> 
#> $Fold8
#> $Fold8$training
#>   [1]   1   2   3   4   5   6   7   8  10  11  12  13  14  15  16  18  19  20
#>  [19]  21  22  24  25  26  27  28  29  30  31  32  33  34  35  37  39  40  42
#>  [37]  43  44  45  46  48  49  50  51  52  53  54  55  56  57  59  60  61  63
#>  [55]  64  65  66  67  68  69  71  72  73  74  76  77  78  79  80  81  82  83
#>  [73]  84  85  86  87  88  89  90  91  92  93  95  97  98 100 101 102 103 104
#>  [91] 105 106 107 108 109 110 111 112 113 114 115 116 117 118 120 121 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold8$test
#>  [1]  47  41  58  36  99  96   9  94  17  38  62 119  23  70  75
#> 
#> 
#> $Fold9
#> $Fold9$training
#>   [1]   2   3   4   5   6   7   8   9  10  11  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  25  26  27  28  29  30  31  32  33  34  36  37  38  40  41
#>  [37]  42  43  46  47  48  49  50  51  52  53  54  55  56  58  59  60  61  62
#>  [55]  63  64  65  66  67  69  70  71  72  73  74  75  76  77  78  79  80  81
#>  [73]  82  83  84  85  86  87  89  90  91  92  93  94  96  98  99 100 101 103
#>  [91] 104 105 106 107 108 109 110 111 112 113 114 116 117 118 119 120 121 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Fold9$test
#>  [1]  24 139  97  68  88 115  35  57  44   1 102  39  12  45  95
#> 
#> 
#> $Fold10
#> $Fold10$training
#>   [1]   1   2   3   4   5   6   8   9  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  23  24  25  26  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  43  44  45  46  47  48  49  50  51  52  53  54  55  57  58  59  60
#>  [55]  61  62  63  64  65  66  68  69  70  72  73  74  75  76  77  78  79  80
#>  [73]  81  84  85  88  89  90  91  92  93  94  95  96  97  98  99 100 101 102
#>  [91] 103 104 105 106 107 108 109 110 111 112 114 115 116 117 118 119 120 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 145 146 147 148 149 150
#> 
#> $Fold10$test
#>  [1]  71  27  42  86  87  10  16 121  83 144  67 113  56  82   7
#> 
#>