Shrinkage-based Diagonal Quadratic Discriminant Analysis (SDQDA)

Given a set of training data, this function builds the Shrinkage-based Diagonal Quadratic Discriminant Analysis (SDQDA) classifier, which is based on the DQDA classifier, often attributed to Dudoit et al. (2002). The DQDA classifier belongs to the family of Naive Bayes classifiers, where the distributions of each class are assumed to be multivariate normal. To improve the estimation of the class variances, Pang et al. (2009) proposed the SDQDA classifier which uses a shrinkage-based estimators of each class covariance matrix.

The SDQDA classifier is a modification to QDA, where the off-diagonal elements of the pooled sample covariance matrix are set to zero. To improve the estimation of the pooled variances, we use a shrinkage method from Pang et al. (2009).

qda_shrink_cov(x, ...)

# S3 method for default
qda_shrink_cov(x, y, prior = NULL, num_alphas = 101, ...)

# S3 method for formula
qda_shrink_cov(formula, data, prior = NULL, num_alphas = 101, ...)

# S3 method for qda_shrink_cov
predict(object, newdata, ...)

Arguments

x
Matrix or data frame containing the training data. The rows are the sample observations, and the columns are the features. Only complete data are retained.

...
additional arguments

y
Vector of class labels for each training observation. Only complete data are retained.

prior
Vector with prior probabilities for each class. If NULL (default), then equal probabilities are used. See details.

num_alphas
the number of values used to find the optimal amount of shrinkage

formula
A formula of the form groups ~ x1 + x2 + ... That is, the response is the grouping factor and the right hand side specifies the (non-factor) discriminators.

data
data frame from which variables specified in formula are preferentially to be taken.

object
Fitted model object

newdata
Matrix or data frame of observations to predict. Each row corresponds to a new observation.

Value

qda_shrink_cov object that contains the trained SDQDA classifier

Details

The DQDA classifier is a modification to the well-known QDA classifier, where the off-diagonal elements of the pooled covariance matrix are assumed to be zero -- the features are assumed to be uncorrelated. Under multivariate normality, the assumption uncorrelated features is equivalent to the assumption of independent features. The feature-independence assumption is a notable attribute of the Naive Bayes classifier family. The benefit of these classifiers is that they are fast and have much fewer parameters to estimate, especially when the number of features is quite large.

The matrix of training observations are given in x. The rows of x contain the sample observations, and the columns contain the features for each training observation.

The vector of class labels given in y are coerced to a factor. The length of y should match the number of rows in x.

An error is thrown if a given class has less than 2 observations because the variance for each feature within a class cannot be estimated with less than 2 observations.

The vector, prior, contains the a priori class membership for each class. If prior is NULL (default), the class membership probabilities are estimated as the sample proportion of observations belonging to each class. Otherwise, prior should be a vector with the same length as the number of classes in y. The prior probabilities should be nonnegative and sum to one.

References

Dudoit, S., Fridlyand, J., & Speed, T. P. (2002). "Comparison of Discrimination Methods for the Classification of Tumors Using Gene Expression Data," Journal of the American Statistical Association, 97, 457, 77-87.

Pang, H., Tong, T., & Zhao, H. (2009). "Shrinkage-based Diagonal Discriminant Analysis and Its Applications in High-Dimensional Data," Biometrics, 65, 4, 1021-1029.

Examples

set.seed(42)
n <- nrow(iris)
train <- sample(seq_len(n), n / 2)
sdqda_out <- qda_shrink_cov(Species ~ ., data = iris[train, ])
predicted <- predict(sdqda_out, iris[-train, -5])$class

sdqda_out2 <- qda_shrink_cov(x = iris[train, -5], y = iris[train, 5])
predicted2 <- predict(sdqda_out2, iris[-train, -5])$class
all.equal(predicted, predicted2)

#> [1] TRUE