Tong et al. (2012)'s Lindley-type Shrunken Mean Estimator

An implementation of the Lindley-type shrunken mean estimator utilized in shrinkage-mean-based diagonal linear discriminant analysis (SmDLDA).

tong_mean_shrinkage(x, r_opt = NULL)

Arguments

x
a matrix with `n` rows and `p` columns.

r_opt
the shrinkage coefficient. If `NULL` (default), we calculate the shrinkage coefficient with the formula given just above Equation 5 on page 533 and denoted by \(\hat{r}_{opt}\). We allow the user to specify an alternative value to investigate better approximations.

Value

vector of length `p` with the shrunken mean estimator

References

Tong, T., Chen, L., and Zhao, H. (2012), "Improved Mean Estimation and Its Application to Diagonal Discriminant Analysis," Bioinformatics, 28, 4, 531-537. http://bioinformatics.oxfordjournals.org/content/28/4/531.long